backend/pyproject.toml
[build-system]
requires = ["setuptools>=69.0"]
build-backend = "setuptools.build_meta"

[project]
name = "pricenavigator-backend"
version = "0.1.0"
description = "FastAPI backend for PriceNavigator"
requires-python = ">=3.13,<3.15"
dependencies = [
  "alembic==1.16.5",
  "fastapi==0.128.8",
  "pydantic-settings==2.11.0",
  "sqlalchemy==2.0.44",
  "uvicorn[standard]==0.38.0",
]

[project.optional-dependencies]
dev = [
  "httpx==0.28.1",
  "pytest==8.4.2",
  "pytest-cov==7.0.0",
  "ruff==0.13.0",
]

[tool.setuptools.packages.find]
include = ["app*"]

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py313"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP", "N", "SIM"]
ignore = ["B008", "E501"]

[tool.ruff.format]
quote-style = "double"
